+++ mountwizzard/modeling/modelStandard.py
@@ -60,9 +60,12 @@ class ModelStandard(ModelBase):
                     self.app.mount.loadBaseModel()
                 keepImages = self.app.ui.checkKeepImages.isChecked()
                 refinePoints = self.runModel('Refinement', self.app.modeling.modelPoints.RefinementPoints, directory, settlingTime, simulation, keepImages)
-                for i in range(0, len(refinePoints)):
-                    refinePoints[i]['index'] += len(self.modelData)
-                self.modelData = self.modelData + refinePoints
+                if self.app.ui.checkKeepRefinement.isChecked():
+                    for i in range(0, len(refinePoints)):
+                        refinePoints[i]['Index'] += len(self.modelData)
+                    self.modelData = self.modelData + refinePoints
+                else:
+                    self.modelData = refinePoints
                 self.modelData = self.app.mount.retrofitMountData(self.modelData)
                 name = directory + '_refinement.dat'
                 if len(self.modelData) > 0: